chore: restore a green gate — explicit lint rules, deterministic type gate
`main` was red on all four CI steps after the reshape merge: 3 failing tests,
2 mypy errors, and 341 ruff findings under ruff 0.16.  None of it was caused by
a code change — the gate itself was underspecified.

Root causes, each fixed rather than silenced:

- **The lint rule set was whatever ruff defaulted to that week.** Ruff 0.16
  widened its default selection, so CI turned red on a linter release date
  rather than on a commit.  `[tool.ruff.lint]` now declares the families
  explicitly and the dev extra bounds ruff/mypy to a minor range, so raising
  the bar is a deliberate commit with its findings fixed in it.

- **The type gate depended on install state.** `mypy morgan_brain` reported
  clean on a box with `[mcp,scale]` installed and five import-not-found errors
  on a `[dev]` runner.  Optional third-party modules now carry an explicit
  `ignore_missing_imports` override.

- **`pip install -e ".[dev]"` could not run the suite CI runs.** The MCP server
  is one of the four shipped surfaces and tests/integration exercises it, as do
  the Qdrant tests; neither package was in `dev`.  `dev` now pulls the runtime
  extras, so the documented install command is sufficient.

- **Two formatters and two mypy configs.** `.pre-commit-config.yaml` still ran
  black + isort + ruff, plus a second mypy with `--no-strict-optional` that
  contradicted the strict one in pyproject, and excluded paths deleted long ago
  (`morgan-rag/`, `scripts/refactor`).  It now mirrors the CI gate exactly.

Real defects the newly-enabled rules surfaced:

- `build_doctor_report` was async but blocking top to bottom — `Path.resolve()`
  plus every `conn.execute()` on the event loop.  Split into a synchronous
  probe body handed to `asyncio.to_thread` and one genuinely async endpoint
  check (ASYNC240).
- `_load_champion_override` swallowed every exception and returned `""`,
  silently dropping the learned preprompt on every turn — the exact failure the
  reshape design called out.  It now logs before falling back (BLE001).
- `RedisStreamsBus.stop()` caught `(CancelledError, Exception)` and passed,
  hiding a failed consume loop.  Cancellation is expected and stays silent;
  anything else is logged (S110).
- 45 naive `datetime()` calls in tests of a bi-temporal store, now tz-aware
  (DTZ001).
- A fake signal store in the scheduling tests assigned `[]` on both branches of
  a conditional and read the field nowhere (RUF034).

`UP042` (`str, Enum` → `StrEnum`) and the ambiguous-unicode rules are ignored
with the reasoning recorded in `pyproject.toml`: the first is a wire-format
change that needs its own tested commit, the second only ever fires on
deliberate typography and on the Cyrillic fixtures that prove the FTS5
tokeniser handles a non-Latin script.

Verified in a clean venv with `pip install -e ".[dev]"` and nothing else:
ruff check (0), ruff format --check (254 clean), mypy strict (112 files, 0),
pytest (782 passed, 8 skipped), bandit -r morgan_brain (0 findings).

+++ morgan-brain/tests/e2e/harness.py
@@ -96,7 +96,7 @@ class BenchReport:
         lat = sorted(self.all_latencies)
         if not lat:
             return 0.0
-        idx = min(len(lat) - 1, int(round((pct / 100.0) * (len(lat) - 1))))
+        idx = min(len(lat) - 1, round((pct / 100.0) * (len(lat) - 1)))
         return lat[idx]
 
     def recall_accuracy(self) -> float:
